project/requirements.txt
Django
python-dotenv
requests
